+++ http-streams/lib/Network/Http/Client.hs
@@ -137,6 +137,7 @@ module Network.Http.Client (
     getStatusCode,
     getStatusMessage,
     getHeader,
+    getHeaderMap,
     debugHandler,
     concatHandler,
     concatHandler',